requirements.txt
fastapi
uvicorn
pymysql
beautifulsoup4
requests
streamlit
pandas
python-dotenv
